+++ app/controllers/properties.js
@@ -1,0 +1,6 @@
+import Ember from "ember";
+import AnchorControllerSupport from "ember-anchor/mixins/controller-support";
+
+export default Ember.Controller.extend(AnchorControllerSupport, {
+  queryParams: ['anchor']
+});
