+++ app/models/class.js
@@ -6,6 +6,8 @@ const {attr, belongsTo} = DS;
 export default DS.Model.extend({
   name: attr(),
   methods: attr(),
+  properties: attr(),
+  events: attr(),
   description: attr(),
   ogDescription: attr(),
   parentClass: belongsTo('class', {async: true, inverse: null}),
