+++ app/styles/_class.scss
@@ -1,0 +1,13 @@
+article.chapter ul {
+  list-style-type: none;
+  -webkit-column-count: 2;
+  column-count: 2;
+  -webkit-column-gap: 1em;
+  column-count: 1em;
+  vertical-align: baseline;
+
+  li {
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+}
