+++ app/styles/app.scss
@@ -3,6 +3,8 @@
 @import "neat";
 @import "./components/all";
 
+@import "class";
+
 body {
   margin: 0;
 }
