+++ app/templates/events.hbs
@@ -1,0 +1,5 @@
+{{ember-anchor a="anchor" anchor=anchor}}
+
+{{#each model.events as |event|}}
+  {{class-field-description type="event" field=event}}
+{{/each}}
