+++ app/templates/properties.hbs
@@ -1,0 +1,5 @@
+{{ember-anchor a="anchor" anchor=anchor}}
+
+{{#each model.properties as |property|}}
+  {{class-field-description type="property" field=property}}
+{{/each}}
